requirements_phth.txt
certifi==2022.12.7
charset-normalizer==2.1.1
clip==1.0
ftfy==6.1.1
idna==3.4
joblib==1.2.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
Pillow==9.3.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
